refactor

+++ src/block_document/bounds.rs
@@ -1,6 +1,6 @@
 pub struct Bounds {
-    pub width: Option<u32>,
-    pub height: Option<u32>,
-    pub x: Option<u32>,
-    pub y: Option<u32>,
+    pub width: Option<f32>,
+    pub height: Option<f32>,
+    pub x: Option<f32>,
+    pub y: Option<f32>,
 }
